Parse CLI for arguments using `clap`
