+++ lib/sudo-cli/Cargo.toml
@@ -5,9 +5,7 @@ edition = "2021"
 license = "Apache-2.0 OR MIT"
 
 [dependencies]
-sudo-common = { path = "../sudo-common" }
 clap = { version = "4.0.32", features = ["derive", "cargo"] }
 
 [dev-dependencies]
-assert_cmd = "2.0.8"
-predicates = "2.1.5"
+pretty_assertions = "1.3.0"
